perf(nexus): prerender every docs page as static JSON and read it on client navigation
Moving between docs pages fetched the next document from /api/docs/page?path=…,
a Worker round trip through D1 that cost 1-2s from CN on every page after the
first. The resolver now sits in server/utils and answers two fronts: the query
route it always had, and /api/docs/page/<locale>/<mode>/<path>.json, which is
path-shaped so nitro can prerender it into a static file (a query string never
becomes a file on Pages — that is what sank af99441e0). The build lists every
document in both locales and both body modes, the docs _headers cache window and
JSON content type cover the prefix, and the client reads the static twin first
with exactly one fallback to the query route for documents outside the list.
