+++ apps/nexus/app/pages/docs/docs-page-performance.test.ts
@@ -64,7 +64,7 @@ const nexusAuth = readFileSync(new URL('../../composables/useNexusAuth.ts', impo
 const authApi = readFileSync(new URL('../../../server/api/auth/[...].ts', import.meta.url), 'utf8')
 const sessionAuthSecret = readFileSync(new URL('../../../server/utils/sessionAuthSecret.ts', import.meta.url), 'utf8')
 const vueDevtoolsApiNoop = readFileSync(new URL('../../utils/vue-devtools-api-noop.ts', import.meta.url), 'utf8')
-const docsPageApi = readFileSync(new URL('../../../server/api/docs/page.get.ts', import.meta.url), 'utf8')
+const docsPageApi = readFileSync(new URL('../../../server/utils/docsPageResolver.ts', import.meta.url), 'utf8')
 const nuxtConfig = readFileSync(new URL('../../../nuxt.config.ts', import.meta.url), 'utf8')
 const packageJson = readFileSync(new URL('../../../package.json', import.meta.url), 'utf8')
 const licensePage = readFileSync(new URL('../license.vue', import.meta.url), 'utf8')
